# swagger-docs/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Service
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Gateway local
  - url: http://gateway-srv:8080
    description: Gateway in-cluster
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Developer API key. Only accepted on /v1/checks and /checks.
        Gateway exchanges it internally for a short-lived JWT via Auth Service.
  schemas:
...
    SavedImage:
      type: object
      required: [image_id, verdict, label, confidence, is_public, uploaded_at]
      properties:
        image_id:
          type: string
        user_id:
          type: string
        s3_key:
          type: string
        verdict:
          type: string
        label:
          type: string
        confidence:
          type: number
          format: float
        model_version:
          type: string
          nullable: true
        is_public:
          type: boolean
        uploaded_at:
          type: string
          format: date-time
        is_reported:
          type: boolean
        url:
          type: string
          format: uri
          description: Presigned S3 URL.
    UpdateImageRequest:
      type: object
      properties:
        is_public:
          type: boolean
          nullable: true
...
paths:
...
  /upload/image:
    post:
      summary: Save image after detection
      tags: [Images]
      description: >
        Save an image with its detection result. Gateway validates the detection_token
        (HS256, verifies user, image SHA256 hash, and expiry) then forwards trusted
        fields to Media Service. The detection_token is never forwarded downstream.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, detection_token]
              properties:
                file:
                  type: string
                  format: binary
                detection_token:
                  type: string
                  description: Token from /v1/checks binding result to image bytes and user.
                is_public:
                  type: boolean
                  description: If true, image is public. Defaults to false.
      responses:
        "201":
          description: Image saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedImage"
        "400":
          description: Invalid or expired detection_token, user/image mismatch
        "401":
          description: Unauthorized
        "413":
          description: Image too large
        "415":
          description: Unsupported format
        "502":
          description: Media service error
        "503":
          description: Media service unavailable
  /images:
    get:
      summary: Get current user's images
      tags: [Images]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: is_public
          schema:
            type: boolean
          description: Filter by public/private visibility.
      responses:
        "200":
          description: List of user's images
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SavedImage"
        "401":
          description: Unauthorized
  /image/{image_id}:
    get:
      summary: Get single image
      tags: [Images]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{1,128}$'
      responses:
        "200":
          description: Image metadata with presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedImage"
        "401":
          description: Unauthorized
        "404":
          description: Image not found or invalid image_id

    patch:
      summary: Update image metadata
      tags: [Images]
      description: Update image visibility. Owner or admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImageRequest"
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedImage"
        "400":
          description: No valid fields to update
        "401":
          description: Unauthorized
        "403":
          description: Not owner and not admin
        "404":
          description: Image not found

    delete:
      summary: Delete image
      tags: [Images]
      description: Delete an image. Owner or admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: image_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image deleted
        "401":
          description: Unauthorized
        "403":
          description: Not owner and not admin
        "404":
          description: Image not found
...
  /admin/models:
    get:
      summary: List all model versions (admin)
      tags: [Models]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model version history
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
...